Change Position struct balance to long long and quantity to int

+++ pe_exchange.h
@@ -15,8 +15,8 @@ typedef struct Position Position;
 
 struct Position {
     char product[MAX_PRODUCT_LEN];
-    unsigned int quantity;
-    int balance;
+    int quantity;
+    long long balance;
 };
